requirements.txt
django-filter
openpyxl
django_extensions
kaleido
plotly
